api/requirements.txt
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Flask-Script==2.0.6
Flask==1.1.2
Jinja2==2.11.2
SQLAlchemy==1.3.17
Werkzeug==1.0.1
alembic==1.4.2
flasgger==0.9.4
itsdangerous==1.1.0
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
setproctitle==1.1.10
six==1.15.0
python-dotenv
bcrypt
pyjwt
flask-cors
